decouple InstanceOfServiceKeyYamlToPhpFactory
