+++ packages/config-transformer/packages/format-switcher/src/Converter/KeyYamlToPhpFactory/ImportsKeyYamlToPhpFactory.php
@@ -16,6 +16,11 @@ use PhpParser\Node\Expr\MethodCall;
 use PhpParser\Node\Expr\Variable;
 use PhpParser\Node\Stmt\Expression;
 
+/**
+ * Handles this part:
+ *
+ * imports: <---
+ */
 final class ImportsKeyYamlToPhpFactory implements KeyYamlToPhpFactoryInterface
 {
     /**
