+++ packages/config-transformer/packages/format-switcher/src/Converter/KeyYamlToPhpFactory/ServicesKeyYamlToPhpFactory.php
@@ -6,15 +6,12 @@ namespace Migrify\ConfigTransformer\FormatSwitcher\Converter\KeyYamlToPhpFactory
 
 use Migrify\ConfigTransformer\FormatSwitcher\Contract\Converter\KeyYamlToPhpFactoryInterface;
 use Migrify\ConfigTransformer\FormatSwitcher\Contract\Converter\ServiceKeyYamlToPhpFactoryInterface;
-use Migrify\ConfigTransformer\FormatSwitcher\Exception\NotImplementedYetException;
-use Migrify\ConfigTransformer\FormatSwitcher\Exception\ShouldNotHappenException;
 use Migrify\ConfigTransformer\FormatSwitcher\PhpParser\NodeFactory\ArgsNodeFactory;
 use Migrify\ConfigTransformer\FormatSwitcher\PhpParser\NodeFactory\CommonNodeFactory;
+use Migrify\ConfigTransformer\FormatSwitcher\PhpParser\NodeFactory\Service\ServiceOptionNodeFactory;
 use Migrify\ConfigTransformer\FormatSwitcher\PhpParser\NodeFactory\ServicesPhpNodeFactory;
 use Migrify\ConfigTransformer\FormatSwitcher\PhpParser\NodeFactory\SingleServicePhpNodeFactory;
-use Migrify\ConfigTransformer\FormatSwitcher\Sorter\YamlArgumentSorter;
 use Migrify\ConfigTransformer\FormatSwitcher\ValueObject\VariableName;
-use PhpParser\BuilderHelpers;
 use PhpParser\Node;
 use PhpParser\Node\Arg;
 use PhpParser\Node\Expr\BinaryOp\Concat;
@@ -22,55 +19,14 @@ use PhpParser\Node\Expr\MethodCall;
 use PhpParser\Node\Expr\Variable;
 use PhpParser\Node\Scalar\String_;
 use PhpParser\Node\Stmt\Expression;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
+/**
+ * Handles this part:
+ *
+ * services: <---
+ */
 final class ServicesKeyYamlToPhpFactory implements KeyYamlToPhpFactoryInterface
 {
-    /**
-     * @var string
-     */
-    private const DECORATION_ON_INVALID = 'decoration_on_invalid';
-
-    /**
-     * @var string
-     */
-    private const DECORATION_INNER_NAME = 'decoration_inner_name';
-
-    /**
-     * @var string
-     */
-    private const DECORATION_PRIORITY = 'decoration_priority';
-
-    /**
-     * @var string
-     */
-    private const ARGUMENTS = 'arguments';
-
-    /**
-     * @var string
-     */
-    private const CALLS = 'calls';
-
-    /**
-     * @var string
-     */
-    private const TAGS = 'tags';
-
-    /**
-     * @var string
-     */
-    private const CONFIGURATOR = 'configurator';
-
-    /**
-     * @var string
-     */
-    private const FACTORY = 'factory';
-
-    /**
-     * @var string
-     */
-    private const INSTANCE_OF = '_instanceof';
-
     /**
      * @var string
      */
@@ -107,11 +63,6 @@ final class ServicesKeyYamlToPhpFactory implements KeyYamlToPhpFactoryInterface
     private $argsNodeFactory;
 
     /**
-     * @var YamlArgumentSorter
-     */
-    private $yamlArgumentSorter;
-
-    /**
      * @var SingleServicePhpNodeFactory
      */
     private $singleServicePhpNodeFactory;
@@ -122,22 +73,27 @@ final class ServicesKeyYamlToPhpFactory implements KeyYamlToPhpFactoryInterface
     private $serviceKeyYamlToPhpFactories = [];
 
     /**
+     * @var ServiceOptionNodeFactory
+     */
+    private $serviceOptionNodeFactory;
+
+    /**
      * @param ServiceKeyYamlToPhpFactoryInterface[] $serviceKeyYamlToPhpFactories
      */
     public function __construct(
         ServicesPhpNodeFactory $servicesPhpNodeFactory,
         CommonNodeFactory $commonNodeFactory,
         ArgsNodeFactory $argsNodeFactory,
-        YamlArgumentSorter $yamlArgumentSorter,
         SingleServicePhpNodeFactory $singleServicePhpNodeFactory,
+        ServiceOptionNodeFactory $serviceOptionNodeFactory,
         array $serviceKeyYamlToPhpFactories
     ) {
         $this->servicesPhpNodeFactory = $servicesPhpNodeFactory;
         $this->commonNodeFactory = $commonNodeFactory;
         $this->argsNodeFactory = $argsNodeFactory;
-        $this->yamlArgumentSorter = $yamlArgumentSorter;
         $this->singleServicePhpNodeFactory = $singleServicePhpNodeFactory;
         $this->serviceKeyYamlToPhpFactories = $serviceKeyYamlToPhpFactories;
+        $this->serviceOptionNodeFactory = $serviceOptionNodeFactory;
     }
 
     public function getKey(): string
@@ -160,12 +116,6 @@ final class ServicesKeyYamlToPhpFactory implements KeyYamlToPhpFactoryInterface
                     $nodes = array_merge($nodes, $freshNodes);
                     continue 2;
                 }
-            }
-
-            if ($serviceKey === self::INSTANCE_OF) {
-                $instanceOfNodes = $this->resolveInstanceOf($serviceValues);
-                $nodes = array_merge($nodes, $instanceOfNodes);
-                continue;
             }
 
             if (isset($serviceValues[self::RESOURCE])) {
@@ -191,111 +141,16 @@ final class ServicesKeyYamlToPhpFactory implements KeyYamlToPhpFactoryInterface
             } else {
                 $args = $this->argsNodeFactory->createFromValues([$serviceKey]);
                 $setMethodCall = new MethodCall(new Variable(VariableName::SERVICES), 'set', $args);
-                $setMethodCall = $this->convertServiceOptionsToNodes($serviceValues, $setMethodCall);
+                $setMethodCall = $this->serviceOptionNodeFactory->convertServiceOptionsToNodes(
+                    $serviceValues,
+                    $setMethodCall
+                );
             }
 
             $nodes[] = new Expression($setMethodCall);
         }
 
         return $nodes;
-    }
-
-    private function convertServiceOptionsToNodes(array $servicesValues, MethodCall $methodCall): MethodCall
-    {
-        foreach ($servicesValues as $serviceConfigKey => $value) {
-            // options started by decoration_<option> are used as options of the method decorate().
-            if (strstr($serviceConfigKey, 'decoration_')) {
-                continue;
-            }
-
-            switch ($serviceConfigKey) {
-                case 'decorates':
-                    $methodCall = $this->createDecorateMethod($servicesValues, $methodCall);
-                    break;
-
-                case 'deprecated':
-                    $methodCall = $this->createDeprecateMethod($value, $methodCall);
-                    break;
-
-                // simple "key: value" options
-                case 'shared':
-                case 'public':
-                    if ($serviceConfigKey === 'public') {
-                        if ($value === false) {
-                            $methodCall = new MethodCall($methodCall, 'private');
-                        } else {
-                            $methodCall = new MethodCall($methodCall, 'public');
-                        }
-                        break;
-                    }
-
-                    throw new NotImplementedYetException();
-
-                case 'bind':
-                case 'autowire':
-                case 'autoconfigure':
-                    $method = $serviceConfigKey;
-                    if ($serviceConfigKey === 'shared') {
-                        $method = 'share';
-                    }
-
-                    $methodCall = new MethodCall($methodCall, $method);
-                    if ($value === false) {
-                        $methodCall->args[] = new Arg($this->commonNodeFactory->createFalse());
-                    }
-
-                    break;
-
-                case self::FACTORY:
-                case self::CONFIGURATOR:
-                    $args = $this->argsNodeFactory->createFromValuesAndWrapInArray($value);
-                    $methodCall = new MethodCall($methodCall, 'factory', $args);
-                    break;
-
-                case self::TAGS:
-                    /** @var mixed[] $value */
-                    if (count($value) === 1 && is_string($value[0])) {
-                        $tagValue = new String_($value[0]);
-                        $methodCall = new MethodCall($methodCall, 'tag', [new Arg($tagValue)]);
-                        break;
-                    }
-
-                    foreach ($value as $singleValue) {
-                        $args = [];
-                        foreach ($singleValue as $singleNestedKey => $singleNestedValue) {
-                            if ($singleNestedKey === 'name') {
-                                $args[] = new Arg(BuilderHelpers::normalizeValue($singleNestedValue));
-                                unset($singleValue[$singleNestedKey]);
-                            }
-                        }
-
-                        $restArgs = $this->argsNodeFactory->createFromValuesAndWrapInArray($singleValue);
-
-                        $args = array_merge($args, $restArgs);
-                        $methodCall = new MethodCall($methodCall, 'tag', $args);
-                    }
-
-                    break;
-
-                case self::CALLS:
-                    $methodCall = $this->singleServicePhpNodeFactory->createCalls($methodCall, $value);
-                    break;
-
-                case self::ARGUMENTS:
-                    $args = $this->argsNodeFactory->createFromValuesAndWrapInArray($value);
-                    $methodCall = new MethodCall($methodCall, 'args', $args);
-
-                    break;
-
-                default:
-                    throw new ShouldNotHappenException(sprintf(
-                        'Unexpected service configuration option: "%s".',
-                        $serviceConfigKey
-                    ));
-            }
-        }
-
-        return $methodCall;
     }
 
     /**
@@ -342,67 +197,12 @@ final class ServicesKeyYamlToPhpFactory implements KeyYamlToPhpFactoryInterface
 
         if (is_array($serviceValues)) {
             /** @var MethodCall $methodCall */
-            $methodCall = $this->convertServiceOptionsToNodes($serviceValues, $methodCall);
+            $methodCall = $this->serviceOptionNodeFactory->convertServiceOptionsToNodes($serviceValues, $methodCall);
         }
 
         $nodes[] = new Expression($methodCall);
 
         return $nodes;
-    }
-
-    private function createDecorateMethod(array $value, MethodCall $methodCall): MethodCall
-    {
-        $arguments = $this->yamlArgumentSorter->sortArgumentsByKeyIfExists($value, [
-            self::DECORATION_INNER_NAME => null,
-            self::DECORATION_PRIORITY => 0,
-            self::DECORATION_ON_INVALID => null,
-        ]);
-
-        if (isset($arguments[self::DECORATION_ON_INVALID])) {
-            $arguments[self::DECORATION_ON_INVALID] = $arguments[self::DECORATION_ON_INVALID] === 'exception'
-                ? $this->commonNodeFactory->createConstFetch(
-                    ContainerInterface::class,
-                    'EXCEPTION_ON_INVALID_REFERENCE'
-                )
-                : $this->commonNodeFactory->createConstFetch(
-                    ContainerInterface::class,
-                    'IGNORE_ON_INVALID_REFERENCE'
-                );
-        }
-
-        // Don't write the next arguments if they are null.
-        if ($arguments[self::DECORATION_ON_INVALID] === null && $arguments[self::DECORATION_PRIORITY] === 0) {
-            unset($arguments[self::DECORATION_ON_INVALID], $arguments[self::DECORATION_PRIORITY]);
-
-            if ($arguments[self::DECORATION_INNER_NAME] === null) {
-                unset($arguments[self::DECORATION_INNER_NAME]);
-            }
-        }
-
-        array_unshift($arguments, $value['decorates']);
-
-        $args = [];
-        foreach ($arguments as $argument) {
-            // is class const refrence
-            $value = BuilderHelpers::normalizeValue($argument);
-            $args[] = new Arg($value);
-        }
-
-        return new MethodCall($methodCall, 'decorate', $args);
-    }
-
-    private function createDeprecateMethod($value, MethodCall $methodCall): MethodCall
-    {
-        // the old, simple format
-        if (! is_array($value)) {
-            $args = $this->argsNodeFactory->createFromValues([$value]);
-        } else {
-            $items = [$value['package'] ?? '', $value['version'] ?? '', $value['message'] ?? ''];
-
-            $args = $this->argsNodeFactory->createFromValues($items);
-        }
-
-        return new MethodCall($methodCall, 'deprecate', $args);
     }
 
     private function isAlias(string $serviceKey, $serviceValues): bool
@@ -424,7 +224,7 @@ final class ServicesKeyYamlToPhpFactory implements KeyYamlToPhpFactoryInterface
 
         unset($serviceValues[self::CLASS_KEY]);
 
-        $setMethodCall = $this->convertServiceOptionsToNodes($serviceValues, $setMethodCall);
+        $setMethodCall = $this->serviceOptionNodeFactory->convertServiceOptionsToNodes($serviceValues, $setMethodCall);
         $nodes[] = new Expression($setMethodCall);
 
         return $nodes;
@@ -445,27 +245,6 @@ final class ServicesKeyYamlToPhpFactory implements KeyYamlToPhpFactoryInterface
         $methodCall->args[] = new Arg(new String_($serviceName));
 
         return $methodCall;
-    }
-
-    /**
-     * @return Node[]
-     */
-    private function resolveInstanceOf($serviceValues): array
-    {
-        $nodes = [];
-        foreach ($serviceValues as $instanceKey => $instanceValues) {
-            $classReference = $this->commonNodeFactory->createClassReference($instanceKey);
-
-            $servicesVariable = new Variable(VariableName::SERVICES);
-            $args = [new Arg($classReference)];
-
-            $instanceofMethodCall = new MethodCall($servicesVariable, 'instanceof', $args);
-            $instanceofMethodCall = $this->convertServiceOptionsToNodes($instanceValues, $instanceofMethodCall);
-
-            $nodes[] = new Expression($instanceofMethodCall);
-        }
-
-        return $nodes;
     }
 
     /**
