+++ phpstan.neon
@@ -43,3 +43,4 @@ parameters:
          - '#Cognitive complexity for "Migrify\\ConfigTransformer\\FormatSwitcher\\Converter\\KeyYamlToPhpFactory\\ServicesKeyYamlToPhpFactory\:\:convertServiceOptionsToNodes\(\)" is \d+, keep it under 8#'
          - '#Cognitive complexity for "Migrify\\ConfigTransformer\\FormatSwitcher\\Converter\\KeyYamlToPhpFactory\\ServicesKeyYamlToPhpFactory\:\:convertYamlToNodes\(\)" is \d+, keep it under 8#'
          - '#Class cognitive complexity for "ServicesKeyYamlToPhpFactory" is \d+, keep it under 50#'
+         - '#Cognitive complexity for "Migrify\\ConfigTransformer\\FormatSwitcher\\PhpParser\\NodeFactory\\Service\\ServiceOptionNodeFactory\:\:convertServiceOptionsToNodes\(\)" is 22, keep it under 8#'
